refactor:优化代码

+++ blog-server-reconstruct/src/main/java/com/aklry/controller/AdminController.java
@@ -68,7 +68,7 @@ public class AdminController {
     @PostMapping("/delete")
     public Result deleteById(@RequestBody Integer id) {
         result = Utils.getResult();
-        adminService.deleteById(id.intValue());
+        adminService.deleteById(id);
 
         result.flag = true;
         result.message = "删除成功";
